Post a deprecation warning for distutils configs
Since we can't do anything about them in the transition (CPython is
dropping support for those entirely), there's nothing we can do but to
tell users to not use them.

This also accounts for Homebrew and Linuxbrew for now. Hopefully they
will come up with better solutions that don't trigger the location
mismatch warning.