Remove dead code
I have tried to remove the .egg-link
(in which case the package is not considered editable),
and removing the package while leaving the .egg-link
(in which case the package is not shown).
I could not produce this "Editable install not found" message
with pip 21.2.

So I think this is dead code and I'm replacing this test by an assert.