Remove unconditionally skipped test (fails if skip removed)
The test was introduced in f9990605dc580ad775a1d40852442040bdfb0ba1 (May
2017) and was unconditionally skipped in
6cf5d7db54ee272fa9af66d45a504d5994693ae4. It remains skipped to this
day.

In commit bd850c079851d5458e4af82921739666545b6170, the patched_file
method was removed/renamed. So even if the skip were removed, it would
fail on the undefined attribute.

This was caught via mypy which emits an error due to the missing method.