Rewrite direct URL reinstallation logic
This entirely rewrites the logic when an incoming to-be-installed
resolved candidate is resolved from a direct URL requirement. The
current logic is:

* Always reinstall on --upgrade or --force-reinstall.
* Always reinstall locally available wheels.
* Always reinstall editables or if an installed editable should be
  changed to be non-editable.
* Always reinstall if local PEP 610 information does not match the
  incoming candidate. This includes cases where the URLs are not
  sufficiently similar, or if the resolved VCS revisions differ.
* Do not reinstall otherwise.

Note that this slightly differs from the proposal raised in previous
discussions, where a local non-PEP 508 path would be reinstalled. This
is due to pip does not actually carry this information to the resolver
and it's not possible to distinguish PEP 508 requirements from bare path
arguments.

The logic does not change how version-specified candidates are
reinstalled.