Optimize provider's is_backtrack_cause by caching resolved names
is_backtrack_cause is called for every requirement on every step of
resolution. As it's implemented currently if an identifier is not the
backtrack cause, the entire list is scanned, invoking the .name
property on every element of the list (and possibly parents). Since all
the provider needs to know is presence or absence of the identifier in
the backtrack causes list stash the resolved names in a set on first
invocation and check for presence/absence in that set.