Pass the session's proxies property to request
After waiting a long time for a fix on the side of the request that could fix pip install issues with proxy as already discussed in #9691. It seems that such changes as mentioned in psf/requests#5735 will take a while to happen, probably only on version 3 of the request. 

Therefore, I'm suggesting a change on the pip side as already proposed in [issuecomment-79051687]https://github.com/pypa/pip/issues/9568#issuecomment-790516875 and [issuecomment-939373868]https://github.com/psf/requests/pull/5735#issuecomment-939373868 

I think it's important that pip address this issue asap since the newer versions of pip are not working properly behind a proxy.