Proposal for exposing `required-by` info in `list` command
cf #10036

The new proposal is to expose the information when the command is
run with `--verbose`, as for `location` or `installer`.

This allows oneliners like:

    pip list --outdated --verbose | awk '{if($7 ~ "roll") print $1}'

Or with more details and a table output:

    pip list --outdated --verbose | awk '{if($7 ~ "roll") print $1,$2,$3}' | column -t

Note: I've not adapted unit tests nor added new ones. Let's see if the
proposal can reach some consensus before :)