Improve warning message for using pip as root
Using PIP as rot is not recommended, however the recommendation
given by PIP in the warning message when someone usess PIP as root
contains recommendeation that might be misleading.

The message only mentiones recommendation of using virtualenv,
however this recommendation does not apply to building containers.

This has been extensively discussed with varying opinions in
the #10556. There are a number of users who use PIP to build
containers who consider using virtualenv as an antipattern,
however there are also PIP maintainers who claim that the
warning message is ok, even if there were attempts to
implement the recommended approach as documented in
https://github.com/apache/airflow/pull/19189 where attempts
to follow the recommendation failed.

This PR attempts to modify the message in a minimal way to still
keep the original recommendation but also provides the container
developers with more appropriate option of switching to another
user and using --user flag instead.