speedup Wheel.support_index_min
Given an input tag list of size `n` and wheel file tags of size `m` this
method is currently `O(n*m)` because it iterates over the set of file tags
then for each file tag it iterates over the input tags.

We can do much better and get `O(m)` time complexity by iterating the input
tags instead, and doing a cheap `O(1)` membership test among the set of file
tags. As a side benefit, this also allows early-termination of the loop.

The impact of this seemingly trivial change is surprisingly big: for a run
of `pip-compile` on macOS, which calls this method many times with large
inputs this changes gives a ~50% speedup on end-to-end `pip-compile` time,
from ~8s down to ~4s!