add --environment-markers option
Add an --environment-markers option to requirements based command which
is used to set the environment markers used when evaluating distribution
requirements during a resolve.

These markers are provided in a JSON file that can be created on the
target using:

python -c "from pip._vendor.packaging.markers import default_environment; print(default_environment())" > env.json

When not specified the default environment markers associated with the
host platform and Python interpreter are used.