Forward 'message' argument of BackendInvalid to Exception
The 'BackendInvalid' exception does not include traceback information like the other backend exceptions, but instead exposes a 'message'.  Since the intention seems to have been that this field would be some piece of useful customer-facing information, it makes sense to forward this up to the Exception base class so that it's rendered on the console. Without this, customers see an opaque `pip._vendor.pep517.wrappers.BackendInvalid` whereas with this customers will see `pip._vendor.pep517.wrappers.BackendInvalid: message` if thrown by the backend.