Load user config with existing `PIP_CONFIG_FILE`
If `PIP_CONFIG_FILE` is set to an existing file, the user configuration
is currently not loaded. This commit changes this behavior, so that the
user configuration will be loaded either way.

This adheres to the
[documentation](https://github.com/pypa/pip/blame/f66b3e8d0168990e35f981e77b6abe6a6974c4ee/docs/html/topics/configuration.md#L84-L105)
and fixes #10643.

Implementing tests for this is currently not possible without more
monkey patching methods, which would have to allow us to load temporary
files for config files with fixed paths.