Add /src explicitly to MYPY_PATH
Newer versions of Mypy seems to prefer packages in the site-packages
directory over those found in the project directory. This causes issues
if Mypy is installed in an environment that has an older pip version
present, e.g. pre-commit. Adding our pip into MYPY_PATH seems to fix
the issue.