Add note to divert away from adding content to user guide
This guide is being broken up, and multiple folks have now tried to add
content to it instead of adding dedicated pages for it.

This note should help direct contributors away from adding more content
on this page, to stop the bleeding and avoid regressing on the amount of
content we'll have to move out of this page later.