Validating that the bin_path executable exists
On Linux, when pip is installed in /usr/local/bin/pip, shutil.which detects this.  The binary_prefix is defaulting to /usr/bin/.  If the file isn't symlinked to /usr/bin it will error out.  Adding in a check to verify that the (binary_prefix, exe_name) exists should be done before passing to samefile to prevent errors from being thrown and for the -m invocation to then be used.