Fix git submodule usage in the testsuite
Git patched a security issue (CVE-2022-39253) which prevents the use of
local file pathed submodules by default. The reason for this is that
untrusted submodules could be abused to expose file system contents.

The pip test suite tests handling of git submodules and to do so creates
a git repo which it then uses as a submodule. This was failing due to
git's stricter rules on acceptable submodules. Thankfully, pip creates
and controls the content of this git repo which means pip can trust it.
To express that to git we pass `-c protocol.file.allow=always` to the
git submodule add command which overrides the default behavior.

This fixes https://github.com/pypa/pip/issues/11540