Add '--ignore-constraint' option
Constraint files are often shared across an organization. When
developing a package included in such a constraint file, it is not
possible to install the package with constraints since the constraint on
the package prevents us installing a development version.

    ❯ cd my-amazing-package
    ❯ cat constraints.txt
    my-amazing-package==1.2.3
    Jinja2==3.1.2
    iso8601==1.1.0
    msgpack==1.0.4
    ❯ pip install -c constraints.txt .
    Processing /dev/my-amazing-package
      Preparing metadata (setup.py) ... done
    ERROR: Cannot install my-amazing-package 1.2.4.dev1 (from /dev/my-amazing-package) because these package versions have conflicting dependencies.

    The conflict is caused by:
        The user requested my-amazing-package 1.2.4.dev1 (from /dev/my-amazing-package)
        The user requested (constraint) my-amazing-package===1.2.4.dev1

    To fix this you could try to:
    1. loosen the range of package versions you've specified
    2. remove package versions to allow pip attempt to solve the dependency conflict

    ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts

Resolve this by allowing users to opt out of individual constraints to
the 'install', 'wheel', and 'download' subcommands. This is rather
manual but it's expected that tools like tox could automatically
generate a value for this option when invoking 'pip install' command.

   ❯ pip install -c constraints.txt --ignore-constraint my-amazing-package .
   ❯ pip wheel -c constraints.txt --ignore-constraint my-amazing-package .
   ❯ pip download -c constraints.txt --ignore-constraint my-amazing-package .

This is only added for the '2020-resolver' resolver, not the
'legacy-resolver' resolver, given the latter is deprecated for removal.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Fixes: #7839