Added --editable-requirements option to install command
When installing a local package in editable mode a link is installed to
the source code so that changes to it take effect immediately. However,
requirements of this package are installed in non-editable mode even if
those requirements also exist locally. This is undesirable when you wish
for changes to the requirement package also take immediate effect.

This commit adds an `--editable-requirements` option which will install
requirements of editable packages in editable mode as well where
possible.