Don't exclude setuptools, distribute & wheel from freeze output on Python 3.12+
Due to the advent of build isolation, it is no longer necessary to install
setuptools and wheel in an environment just to install other packages.
Moreover, on Python 3.12 both ensurepip [1] and virtualenv [2] are to stop
installing setuptools & wheel by default. This means that when those packages
are present in a Python 3.12+ environment, it is reasonable to assume that
they are runtime dependencies of the user's project, and therefore should be
included in freeze output.

distribute is just obsolete.

[1] https://github.com/python/cpython/issues/95299
[2] https://github.com/pypa/virtualenv/pull/2558