Allow pip install --executable=/path to set #! path in scripts.
This is useful for cross-compiling, when the path to the Python
executable that is being used to run pip, available at
sys.executable, is not the same as the path to the Python executable
that the installed product in a cross-compiled system needs to run.

For example, if your toolchain lives at /build/aarch64--netbsd but
you're building software to be installed at /opt/pkg, then
sys.executable might be /build/aarch64--netbsd/bin/python3.10 but the
shebang line of any scripts installed must /opt/pkg/bin/python3.10.
With this change, you can accomplish that with:

/build/aarch64--netbsd/bin/python3.10 -m pip install --executable=/opt/pkg/bin/python3.10 ...

fix https://github.com/pypa/pip/issues/12087