Correct typo in 23.2 Changelog Bug Fixes
The first entry in the bug fixes section for the 23.2 release's
changelog incorrectly listed the the PR/Issue for "Fix pip 
completion --zsh" as #11416. However, the PR that fixed this
issue was actually #11417. This commit fixes this oversight/off by one error so that the changelog points to the correct PR.