Replace python2 deprecation with a badge of supported python versions
The python world has (mostly) moved on from Python 2. Anyone not already aware of the py2->py3 migration is probably new to the ecosystem and started on Python 3.

Additionally, it's convenient to see at a glance what versions of Python are supported by the current release. This pulls from PyPI versions, so will not immediately match `main` (we can probably change this to match `main` if preferred).

So by doing this it's both more useful going forward, and also lets us drop the explicit notice about dropping Python 2 support.