Upgrade mypy
Mostly straightforward.

The new message.get_all type ignores out of fear it may return None are
minorly annoying. You can see some of the extra errors on the pip
codebase in the mypy_primer of the typeshed PR that made the change
https://github.com/python/typeshed/pull/9620

The abstract errors are because FakeDistribution doesn't actually
implement any of the Protocol it claims to implement