Use strict optional checking in legacy resolver
The assert in _set_req_to_reinstall is definitely correct, the three
call sites have guards

The assert in _populate_link is a little trickier. It's not obvious to
me how to prove it will never trigger. Note that if link were ever None,
it could potentially cause AttributeError in Cache._get_cache_path_parts
and direct_url_from_link

The assert in _resolve_one is safe, if req_to_install.name was None it
would raise TypeError in canonicalize_name

I can't prove that req.name is not None in get_installation_order
and the code would work fine if that were the case (since it's a
defaultdict), so I opted to just ignore