The purpose of this change is to avoid useless HTTP(S) requests
In essence, if the cache already contains the file we want to
get over HTTPS *and* it matches the hash that we have at hand,
there is no use checking for different/newer versions over the
network.

A use case for this is if you have a requirements.txt that contains
packages with urls and hashes, if you reinstall packages, you do
not need to go over the network.