Stop injecting `wheel` as a build dep fallback
PEP 517 doesn't mandate depending on `wheel` when a `__legacy__` setuptools
fallback is used. Historically, it used to be assumed as necessary, but
later it turned out to be wrong. The reason is that `setuptools`'
`get_requires_for_build_wheel()` hook already injects this dependency when
building wheels is requested [[1]]. It also used to have this hint in
the docs, but it was corrected earlier [[2]].

It could be argued that this is an optimization as `pip` will request
building wheels anyway. However, it also shows up in the docs, giving
the readers a wrong impression of what to put into
`[build-system].requires` when they start a new project using setuptools.

This patch removes `wheel` from said `requires` list fallback in the
docs and the actual runtime.

[1]: https://github.com/pypa/setuptools/blob/v40.8.0/setuptools/build_meta.py#L130
[2]: https://github.com/pypa/setuptools/pull/3056