Change Python version support to match CPython
Currently, we are supporting Python version until their usage falls
low-enough. This worked well, when we had fewer users and Python's
sliding window moved slower.

We have, however, not used the threshold for the last few Python
releases and our metrics for determining usage levels are not
particularly robust against being overrepresentative of certain usage
patterns (we're seeing this in the current usage numbers).

Aligning with CPython helps ensure that the maintainers of pip aren't
subsidising the cost of EOL Python versions indefinitely.