Strip RIGHT-TO-LEFT OVERRIDE and LEFT-TO-RIGHT OVERRIDE characters from AUTHORS.txt
Since we don't know anything about the names, we assume anything in between
RIGHT-TO-LEFT OVERRIDE and LEFT-TO-RIGHT OVERRIDE (or the end of the name)
should be spelled backwards.

This resulted in a duplicate author name,
because it uses different Unicode form.
So I also added a call to unicodedata.normalize.

Fixes https://github.com/pypa/pip/issues/12467