Remove Link's dependency on KeyBasedCompareMixin
Unfortunately due to dataclasses' poor handling of slots under 3.10,
converting the class to a dataclass is not an option. I would get
these errors which would be nontrivial to work around.

ValueError: '_hashes' in __slots__ conflicts with class variable

However, this is the last use of KeyBasedCompareMixin so after this
commit, it can be removed entirely!