Replace tenacity.retry with our own decorator
Tenacity is a 1600+ LOC dependency which is only used for a convenient
`@retry` decorator... which is used a grand total of two times. This is a
lot of code for what can be trivially reimplemented by us.

This also incidentally improves startup performance as tenacity imports
asyncio which is expensive and unnecessary for us.