Cache supported tags in resolver factory
I've observed get_supported() consume up to 10% of the installation step
when installing many packages. Each call can take 1-5ms (presumably only
on Linux due to the large number of supported tags) and there is one
call for every lookup in the cache. As all of these calls are made with
no arguments, the tags can be trivially queried in advance during
factory initialization.