Support completion of installation extras in zsh
When using the `pip install '.[extra,<tab>' ` in zsh, this should list
the currently available extras if present in pyproject.toml.

I'm not a big expert in zsh, so there are some limitations, in particular
one must use quotes (', or "), otherwise, the completer split the line
on `[`, and we don't get called.

The completion does redisplay the full, already types extras, so
`.[dev,t<tab>` will suggest things like `.[dev,tests]`, `.[dev,tox]`
instead of just `test`, `tox`.

Test not yet icluded, but I can add if there is some interest.
If you do not think this has a chance to get in, please let me know and
I can close, I know you are fairly busy, and can understand this is
extra maintenance.

It currently only support cwd for pyproject.toml, and does not suggest
to add extras if a `[` is opened.